# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Lopez Calculator API",
    "description" : "This is a simple API",
    "contact" : {
      "email" : "al08583@georgiasouthern.edu"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "2.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/alexalpz/LopezCalculatorAPI/2.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "https://alexalopez.xyz/?id={num1}&id={num2}&id={add}&id={subtract}&id={multiply}&id={divide}&id={power}",
    "description" : "My public calculator application. "
  } ],
  "tags" : [ {
    "name" : "public",
    "description" : "Accessible to the public"
  }, {
    "name" : "admin",
    "description" : "Accessible to admin only"
  } ],
  "paths" : {
    "/" : {
      "get" : {
        "tags" : [ "public" ],
        "summary" : "Searches values",
        "description" : "By passing in the appropriate options, you can search for\navailable inventory in the system\n",
        "operationId" : "searchNumber",
        "parameters" : [ {
          "name" : "num1",
          "in" : "query",
          "description" : "Passing first number",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "num2",
          "in" : "query",
          "description" : "Passing second number",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "name" : "add",
          "in" : "query",
          "description" : "Adds two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "subtract",
          "in" : "query",
          "description" : "Subtracts two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "multiply",
          "in" : "query",
          "description" : "Multiplies two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "divide",
          "in" : "query",
          "description" : "Divides two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "power",
          "in" : "query",
          "description" : "Raises power of two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/calculator"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      },
      "put" : {
        "tags" : [ "public" ],
        "summary" : "Updates values",
        "description" : "By passing in the appropriate options, you can search for\navailable inventory in the system\n",
        "operationId" : "updateNumber",
        "parameters" : [ {
          "name" : "num1",
          "in" : "query",
          "description" : "Passing first number",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "num2",
          "in" : "query",
          "description" : "Passing second number",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "add",
          "in" : "query",
          "description" : "Adds two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "subtract",
          "in" : "query",
          "description" : "Subtracts two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "multiply",
          "in" : "query",
          "description" : "Multiplies two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "divide",
          "in" : "query",
          "description" : "Divides two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "power",
          "in" : "query",
          "description" : "Raises power of two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/calculator"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      },
      "post" : {
        "tags" : [ "admin" ],
        "summary" : "Creates a resource",
        "description" : "Adds an item to the system",
        "operationId" : "addsNumber",
        "parameters" : [ {
          "name" : "num1",
          "in" : "query",
          "description" : "Passing first number",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "num2",
          "in" : "query",
          "description" : "Passing second number",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "add",
          "in" : "query",
          "description" : "Adds two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "subtract",
          "in" : "query",
          "description" : "Subtracts two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "multiply",
          "in" : "query",
          "description" : "Multiplies two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "divide",
          "in" : "query",
          "description" : "Divides two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "power",
          "in" : "query",
          "description" : "Raises power of two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/calculator"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      },
      "delete" : {
        "tags" : [ "admin" ],
        "summary" : "Deletes values",
        "description" : "By passing in the appropriate options, you can search for\navailable inventory in the system\n",
        "operationId" : "deletesNumber",
        "parameters" : [ {
          "name" : "num1",
          "in" : "query",
          "description" : "Passing first number",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "num2",
          "in" : "query",
          "description" : "Passing second number",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "add",
          "in" : "query",
          "description" : "Adds two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "subtract",
          "in" : "query",
          "description" : "Subtracts two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "multiply",
          "in" : "query",
          "description" : "Multiplies two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "divide",
          "in" : "query",
          "description" : "Divides two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        }, {
          "name" : "power",
          "in" : "query",
          "description" : "Raises power of two integers",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/calculator"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "calculator" : {
        "description" : "My simple calculator application"
      }
    }
  }
}